Added Graphical overview of translation progress
It's a per view live rendered picture of our Transifex translation progress with all added languages.

I got this Idea while I read the Owncloud Readme.md. They also use Transifex so I thought it's maybe a service of Transifex. Well I haven't found it on the official Docs but changing the link from OC to the Ampache-equivalent worked.. so , here it is.

:-P